+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [ Unreleased ] - ReleaseDate
+## [ 0.11.4 ] - 2023-03-26
 
 ### Fixed